worker/requirements.txt
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
docker==6.1.3
pydantic==2.5.0
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3